requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
gunicorn==21.2.0
openpyxl==3.1.2
bcrypt==4.1.2
PyJWT==2.8.0
email-validator==2.1.0
WTForms==3.1.1
python-dateutil==2.8.2 